ajusta tipo de dado de total_products e total para DECIMAL(10, 2) nas migrações; atualiza seeders com valores correspondentes; implementa funções utilitárias para agregação de dados em relatórios e validação de filtros de data

+++ backend/src/database/migrations/04-create-sales.ts
@@ -43,11 +43,11 @@ export default {
       },
       totalProducts: {
         field: 'total_products',
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(10, 2),
         allowNull: false,
       },
       total: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(10, 2),
         allowNull: false,
       },
     });
